warning対応
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@1403 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Samples/BoxStack/main.cpp
@@ -238,10 +238,10 @@ void keyboard(unsigned char key, int x, int y){
 void timer(int id){
 	/// 時刻のチェックと画面の更新を行う
 	float DT = 0.05f;
-	int times = DT / scene->GetTimeStep();
+	int times = (int)(DT / scene->GetTimeStep());
 	for(int i=0; i<times; ++i) scene->Step();
 	glutPostRedisplay();
-	glutTimerFunc(1000*DT, timer, 0);
+	glutTimerFunc((unsigned int)1000*DT, timer, 0);
 }
 
 /**
